refactor: 공고 등록 기능 리팩토링
